requirements.txt
pandas
mcp[cli]
peewee
fake
pytest
uv
psycopg2-binary
mcp-client-cli